backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.5
openai==1.3.8
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
twilio==8.10.0
python-dateutil==2.8.2
